requirements.txt
celery==4.3.0
Django==2.2.7
djangorestframework==3.10.3
ipdb==0.12.2
requests==2.22.0